requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dateutil>=2.8.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-telegram-bot>=20.7
pytz>=2023.3
APScheduler>=3.10.4
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
SQLAlchemy>=2.0.0
openpyxl>=3.1.2
reportlab>=4.0.4
